+++ Sources/CyPlayer.h
@@ -164,7 +164,7 @@ public:
 	int calculateUnitSupply();
 	int64_t calculatePreInflatedCosts();
 	int calculateInflationRate();
-	int calculateInflatedCosts();
+	int64_t calculateInflatedCosts();
 
 	int calculateGoldRate();
 	int calculateTotalCommerce();
